🎨 (structure): put components in subfolders
