+++ apps/office-booker/src/app/bookings/map-bookings/map-bookings.component.html
@@ -26,3 +26,4 @@
     </mat-grid-list>
   </ng-template>
 </div>
+
