+++ apps/office-booker/src/app/bookings/personal-bookings/personal-bookings.component.html
@@ -1,0 +1,1 @@
+<p>personal-bookings works!</p>
