+++ apps/office-booker/src/app/bookings/personal-bookings/personal-bookings.component.ts
@@ -1,0 +1,13 @@
+import { Component} from '@angular/core';
+
+@Component({
+  selector: 'office-booker-personal-bookings',
+  templateUrl: './personal-bookings.component.html',
+  styleUrls: ['./personal-bookings.component.css'],
+})
+export class PersonalBookingsComponent{
+  constructor() {
+    //
+  }
+
+}
