+++ apps/office-booker/src/app/shared/menu-bar/menu-bar.component.ts
